+++ BACKEND/doctor_service/src/main/java/com/hospital/doctor_service/cotroller/DoctorController.java
@@ -3,6 +3,7 @@ package com.hospital.doctor_service.cotroller;
 import com.hospital.doctor_service.dto.DoctorResponseDTO;
 import com.hospital.doctor_service.model.Doctor;
 import com.hospital.doctor_service.service.DoctorService;
+import com.hospital.doctor_service.util.RoleUtil;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,7 +23,12 @@ public class DoctorController {
 
     // Create a new doctor
     @PostMapping
-    public ResponseEntity<DoctorResponseDTO> createDoctor(@RequestBody Doctor doctor) {
+    public ResponseEntity<DoctorResponseDTO> createDoctor(
+            @RequestBody Doctor doctor,
+            @RequestHeader("role") String role) {
+
+        RoleUtil.checkAdmin(role);
+
         return ResponseEntity.ok(doctorService.createDoctor(doctor));
     }
 
@@ -40,15 +46,26 @@ public class DoctorController {
 
     // Update a doctor by ID
     @PutMapping("/{id}")
-    public ResponseEntity<Doctor> updateDoctor(@PathVariable Long id,
-                                               @RequestBody Doctor doctor) {
+    public ResponseEntity<Doctor> updateDoctor(
+            @PathVariable Long id,
+            @RequestBody Doctor doctor,
+            @RequestHeader("role") String role) {
+
+        RoleUtil.checkAdmin(role);
+
         return ResponseEntity.ok(doctorService.updateDoctor(id, doctor));
     }
 
     // Delete a doctor by ID
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteDoctor(@PathVariable Long id) {
+    public ResponseEntity<String> deleteDoctor(
+            @PathVariable Long id,
+            @RequestHeader("role") String role) {
+
+        RoleUtil.checkAdmin(role);
+
         doctorService.deleteDoctor(id);
+
         return ResponseEntity.ok("Doctor deleted successfully");
     }
 }
