+++ BACKEND/feedback-service/src/main/java/com/hospital/feedback_service/controller/FeedbackController.java
@@ -3,6 +3,7 @@ package com.hospital.feedback_service.controller;
 import com.hospital.feedback_service.dto.FeedbackResponseDTO;
 import com.hospital.feedback_service.model.Feedback;
 import com.hospital.feedback_service.service.FeedbackService;
+import com.hospital.feedback_service.util.RoleUtil;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -38,13 +39,24 @@ public class FeedbackController {
 
 
     @PutMapping("/{id}")
-    public Feedback update(@PathVariable Long id, @RequestBody Feedback feedback) {
+    public Feedback update(@PathVariable Long id,
+                           @RequestBody Feedback feedback,
+                           @RequestHeader("role") String role) {
+
+        RoleUtil.checkAdmin(role);
+
         return service.updateFeedback(id, feedback);
     }
 
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Long id,
+                       @RequestHeader("role") String role) {
+
+        RoleUtil.checkAdmin(role);
+
         service.deleteFeedback(id);
     }
+
+
 }
